refactor(app): create shared PlanetDetails model
