+++ lib/app/shared/utils/data.dart
@@ -1,18 +1,4 @@
-import 'package:flutter/foundation.dart';
-
-class PlanetDetails {
-  final String description;
-  final String image;
-  final String name;
-  final bool saved;
-
-  PlanetDetails({
-    @required this.description,
-    @required this.image,
-    @required this.name,
-    @required this.saved,
-  });
-}
+import 'package:solar_system_umpontoseis/app/shared/models/planet_details.dart';
 
 class AppPlanets {
   static PlanetDetails mars = PlanetDetails(
